Starting work on new map editor movement system
Starting work on a new movement system for the map editor.
Basically im attempting to copy how the unity editor does movement/scaling and rotating in the future, but I have NO clue how to handle rotating yet.

I also changed the scenes to an enum for future proofing incase the indexes mess up, we dont have to search through the code and change them. Simply just change the index in the GameScenes enum

+++ Assets/MenuScripts/MenuController.cs
@@ -44,7 +44,7 @@ public class MenuController : MonoBehaviour
         }
 
         string robotNameToLoad = PlayerPrefs.GetString("lastRobot", robots[0].robotName);
-        int levelIdToLoad = PlayerPrefs.GetInt("lastLevel", levels[0].levelId);
+        int levelIdToLoad = PlayerPrefs.GetInt("lastLevel", (int)levels[0].levelId);
 
         // Generate level toggles
         foreach (LevelScriptableObject level in levels)
@@ -58,7 +58,7 @@ public class MenuController : MonoBehaviour
             });
             toggle.group = togglesGroup;
 
-            if (level.levelId == levelIdToLoad)
+            if ((int)level.levelId == levelIdToLoad)
             {
                 activeLevel = level;
                 toggle.isOn = true;
@@ -96,7 +96,7 @@ public class MenuController : MonoBehaviour
     public void SelectLevel(LevelScriptableObject activeLevel)
     {
         MenuController.activeLevel = activeLevel;
-        PlayerPrefs.SetInt("lastLevel", activeLevel.levelId);
+        PlayerPrefs.SetInt("lastLevel", (int)activeLevel.levelId);
     }
 
     public void SelectRobot(RobotScriptableObject activeRobot)
@@ -163,6 +163,6 @@ public class MenuController : MonoBehaviour
 
     public void PlaySim()
     {
-        SceneManager.LoadScene(activeLevel.levelId);
+        SceneManager.LoadScene((int)activeLevel.levelId);
     }
 }
